Adding random to the repository as a subfolder
Removed random from submodule because it created more headaches
than those it solved. Composer couldn't download it and
downloading from github as a zip meant you were getting an
nlp-tools package that didn't work. It is important to be
able to download it as a simple zip file, extract and go.